refactor(server): organize document services
