+++ packages/server/src/__tests__/attachments-route.test.ts
@@ -16,7 +16,7 @@ import {
 } from "../services/attachment.js";
 import { MemoryAttachmentBlobStore } from "../services/attachment-blob-store.js";
 import { editMessage, lockFileAttachmentRefsIfPresent, sendMessage } from "../services/chat/message.js";
-import { validateMessageAttachmentRefs } from "../services/doc-snapshots.js";
+import { validateMessageAttachmentRefs } from "../services/chat/message-attachment-validation.js";
 import { ensureMembership } from "../services/team/membership.js";
 import { uuidv7 } from "../uuid.js";
 import { createAdminContext, createTestAdmin, useTestApp } from "./helpers.js";
