+++ packages/server/src/__tests__/documents.test.ts
@@ -8,8 +8,8 @@ import { members } from "../db/schema/members.js";
 import { organizations } from "../db/schema/organizations.js";
 import { users } from "../db/schema/users.js";
 import { createAgent } from "../services/agents/identity.js";
-import { docAuthorForAgentUuid } from "../services/doc-author.js";
-import { createComment } from "../services/document.js";
+import { docAuthorForAgentUuid } from "../services/document-review/author.js";
+import { createComment } from "../services/document-review/document.js";
 import { uuidv7 } from "../uuid.js";
 import { createAdminContext, createTestAgent, INVALID_BCRYPT_PLACEHOLDER, useTestApp } from "./helpers.js";
